feat: share detail stacks across tab segments
