+++ apps/akari/app/(tabs)/(index,search,notifications,bookmarks,messages,profile,settings)/messages/_layout.tsx
@@ -1,0 +1,11 @@
+import { Stack } from "expo-router";
+
+export default function MessagesLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="index" />
+      <Stack.Screen name="pending" />
+      <Stack.Screen name="[handle]" />
+    </Stack>
+  );
+}
