+++ apps/akari/app/(tabs)/(index,search,notifications,bookmarks,messages,profile,settings)/post/_layout.tsx
@@ -1,0 +1,9 @@
+import { Stack } from 'expo-router';
+
+export default function PostLayout() {
+  return (
+    <Stack>
+      <Stack.Screen name="[id]" />
+    </Stack>
+  );
+}
