+++ apps/akari/app/(tabs)/(index,search,notifications,bookmarks,messages,profile,settings)/profile/_layout.tsx
@@ -1,0 +1,10 @@
+import { Stack } from 'expo-router';
+
+export default function ProfileLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="index" />
+      <Stack.Screen name="[handle]" />
+    </Stack>
+  );
+}
