+++ apps/akari/app/(tabs)/profile/_layout.tsx
@@ -1,10 +1,1 @@
-import { Stack } from 'expo-router';
-
-export default function ProfileLayout() {
-  return (
-    <Stack screenOptions={{ headerShown: false }}>
-      <Stack.Screen name="index" />
-      <Stack.Screen name="[handle]" />
-    </Stack>
-  );
-}
+export { default } from '../(index,search,notifications,bookmarks,messages,profile,settings)/profile/_layout';
